third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFEPointLightElement-svgdom-z-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'z' property of the SVGFEPointLightElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS pointLight.z.baseVal is 100
PASS pointLight.z.baseVal is 30
PASS successfullyParsed is true

TEST COMPLETE
